requirements.txt
flask
pandas
werkzeug
pyopenssl
python-dotenv